Contig#getNumberOfReads() now returns a long
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1436 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/ace/IndexedAceFileContig.java
@@ -81,7 +81,7 @@ final class IndexedAceFileContig implements AceContig{
 	}
 
 	@Override
-	public int getNumberOfReads() {
+	public long getNumberOfReads() {
 		return readInfoMap.size();
 	}
